+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Move Deployments tab to the second position on the service entity page.
+
 ## [0.12.2] - 2023-12-14
 
 ### Changed