+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 2024-01-10
+
 ### Changed
 
 - Move Deployments tab to the second position on the service entity page.
@@ -285,7 +287,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disable anonymous access.
 
 
-[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.12.2...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/giantswarm/backstage/compare/v0.12.2...v0.13.0
 [0.12.2]: https://github.com/giantswarm/backstage/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/giantswarm/backstage/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/giantswarm/backstage/compare/v0.11.0...v0.12.0